+++ src/ClassBased/components/TodoList.js
@@ -1,5 +1,5 @@
 import { Component } from "react";
-import TodoItem from "./TodoItem";
+import TodoItem from "../FunctionBased/TodoItem";
 
 class TodoList extends Component {
   render() {
