+++ src/FunctionBased/App.css
@@ -17,7 +17,7 @@ body {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  max-width: 600px;
+  max-width: 800px;
   margin: 0 auto;
 }
