+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import TodoContainer from "./components/TodoContainer";
-import "./App.css"
+import TodoContainer from "./FunctionBased/TodoContainer";
+import "./FunctionBased/App.css"
 
 ReactDOM.render(
   <React.StrictMode>
